// frontend/mrsexam-ui/src/api/generated/models/FormPraveen.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FormPraveen = {
    rollNo: number;
    regionNo: number;
    sessionNo: number;
    name?: string | null;
    formCentre?: string | null;
    examCentre?: string | null;
    marks1?: string | null;
    marks2?: string | null;
    marks3?: string | null;
    marks4?: string | null;
    oralMarks?: string | null;
    regLang?: string | null;
    totMarks?: number | null;
    grade?: string | null;
    undFlag?: string | null;
    undClass?: string | null;
    specialRank?: string | null;
    result?: string | null;
    dummyCentreCode?: number | null;
    userId?: string | null;
};